refactor: reorganize component imports and rename LabList to ExpList
