+++ src/app/lab/page.tsx
@@ -1,4 +1,4 @@
-import { LabList } from '@/components/layout'
+import { ExpList } from '@/components/layout'
 import { generateSEO } from '@/config/site'
 import { experiments } from '@/lib/content'
 import React from 'react'
@@ -18,5 +18,5 @@ interface LabPageProps {
 export default async function lab(props: LabPageProps) {
     const searchParams = await props.searchParams;
     const query = searchParams?.query || '';
-    return <LabList experiments={experiments} query={query} />
+    return <ExpList experiments={experiments} query={query} />
 }
